+++ hw_interface/CMakeLists.txt
@@ -1,0 +1,52 @@
+cmake_minimum_required(VERSION 2.8.3)
+project(hw_interface)
+
+##################
+## Dependencies ##
+##################
+add_compile_options(-std=c++11)
+
+set(PKG_DEPS
+)
+
+find_package(catkin REQUIRED COMPONENTS ${PKG_DEPS})
+
+set(LSOC_LIBRARIES "-lsoc")
+
+############
+## catkin ##
+############
+
+catkin_package(
+    CATKIN_DEPENDS ${PKG_DEPS}
+    LIBRARIES ${PROJECT_NAME}
+    INCLUDE_DIRS include
+)
+
+###########
+## Build ##
+###########
+
+include_directories(
+    include
+    ${catkin_INCLUDE_DIRS}
+)
+
+add_library(${PROJECT_NAME} src/gpio.cpp)
+
+#############
+## Install ##
+#############
+
+install(TARGETS ${PROJECT_NAME}
+    LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
+)
+
+install(DIRECTORY include/${PROJECT_NAME}
+   DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
+   FILES_MATCHING PATTERN "*.h" PATTERN "*.hpp"
+)
+
+install(FILES README.md
+    DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
+)
