+++ hw_interface/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>suction_pump</name>
+  <name>hw_interface</name>
   <version>0.0.1</version>
   <description>
    Suction pump interface. Supports schmaltz FXP pump.
@@ -13,9 +13,6 @@
   <buildtool_depend>catkin</buildtool_depend>
   
   <depend>roscpp</depend>
-  <depend>actionlib</depend>
-  <depend>actionlib_msgs</depend>
-  <depend>hw_interface</depend>
   <!-- <depend>libsoc</depend> -->
 
   <exec_depend>message_generation</exec_depend>
